Release 1.1.0 -- Adds optionally-enablable HTTP server.
- Note: your API Key is still required for all interactions!
- The utility of an HTTPS server on something bound to just localhost is
a little dubious, and after consulting with a colleague of mine, he
confirmed my beliefs that as long as the service was not exposed to the
internet in any way, there was no likely possiblity of connection
snooping.  There are no avenues for an XSS attack in any case given that
this is using header-based auth.
- Additionally, Chrome will not allow you to connect to sites with
self-signed certificates unless you upload your certificate as a valid
Certificate Authority.  Although _I_ know that's safe, I imagine that's
even scarier for a user than just running on HTTP alone; so, this will
reduce friction at the expense of only very theoretical security.